requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
pandas>=2.1.4
openpyxl>=3.1.5
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
pytest==8.0.0
pytest-asyncio==0.23.5
openai==1.45.0
email-validator==2.1.0
pdfkit==1.0.0
jinja2==3.1.3
python-dotenv==1.0.1
psycopg2-binary>=2.9.9
